+++ ACP3/Core/Modules/Helper/Action.php
@@ -173,32 +173,6 @@ class Action
      * @param callable $callback
      * @param null|string $path
      *
-     * @return string|array|\Symfony\Component\HttpFoundation\JsonResponse|\Symfony\Component\HttpFoundation\RedirectResponse
-     *
-     * @deprecated since 4.4.4, to be removed with version 4.5.0
-     */
-    public function handleCreatePostAction(callable $callback, $path = null)
-    {
-        return $this->handleSaveAction($callback, $path);
-    }
-
-    /**
-     * @param callable $callback
-     * @param null|string $path
-     *
-     * @return string|array|\Symfony\Component\HttpFoundation\JsonResponse|\Symfony\Component\HttpFoundation\RedirectResponse
-     *
-     * @deprecated since 4.4.4, to be removed with version 4.5.0
-     */
-    public function handleEditPostAction(callable $callback, $path = null)
-    {
-        return $this->handleSaveAction($callback, $path);
-    }
-
-    /**
-     * @param callable $callback
-     * @param null|string $path
-     *
      * @return array|string|JsonResponse|RedirectResponse
      */
     public function handleSaveAction(callable $callback, $path = null)